feat: Enable highlighting of color variables in lsp for flutter >= 2.10
This commit modifies the `flutter-tools_plugin.lua` file to enable the highlighting of color variables in the language server protocol (lsp) for Flutter versions 2.10 and above. The `enabled` and `background` options are set to `true`, and the `foreground` option is set to `true`. This change improves the visibility and usability of color variables in the code.