chore: Update docker-image.yml
This commit updates the docker-image.yml file to remove the platforms linux/amd64v2 from the build and push action. The linux/amd64v2 platform is no longer needed and removing it simplifies the configuration. The push option is set to true, and the tags are set to user/app:latest. This change ensures that the latest version of the docker image is built and pushed with the specified tags.